render-requirements.txt
# Render-optimized requirements for RAG system
# Removed some packages that might cause issues on Render's free tier

google-api-python-client==2.108.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
openai==1.3.8
langchain==0.0.350
langchain-openai==0.0.2
chromadb==0.4.18
streamlit==1.28.2
python-dotenv==1.0.0
numpy==1.24.3
pandas==2.1.4
tiktoken==0.5.2
beautifulsoup4==4.12.2
requests==2.31.0
tenacity==8.2.3
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
psycopg2-binary==2.9.9
redis==5.0.1
gunicorn==21.2.0
aiofiles==23.2.1
httpx==0.25.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4